refactored validation for token
