+++ src/form/validation/asset-amount.ts
@@ -1,14 +1,12 @@
 import { BigNumber } from 'bignumber.js';
 import { object } from 'yup';
 
-import { tokenEqualityFn } from '../../components/token-dropdown/token-equality-fn';
-import { emptyTezosLikeToken, TokenInterface } from '../../token/interfaces/token.interface';
-import { isDefined } from '../../utils/is-defined';
+import { assetValidation } from './asset';
 import { bigNumberValidation } from './big-number';
 import { makeRequiredErrorMessage } from './messages';
 
 export const assetAmountValidation = object().shape({
-  asset: object().shape({}).required(makeRequiredErrorMessage('Asset')),
+  asset: assetValidation,
   amount: bigNumberValidation
     .clone()
     .required(makeRequiredErrorMessage('Amount'))
@@ -21,16 +19,6 @@ export const assetAmountValidation = object().shape({
     })
 });
 
-export const assetValidation = object().shape({
-  asset: object()
-    .shape({})
-    .required(makeRequiredErrorMessage('Token'))
-    .test('is-equal', 'Token must be selected', (value: TokenInterface) => {
-      const isEqual = tokenEqualityFn(value, emptyTezosLikeToken);
-      if (isEqual || !isDefined(value.address) || !isDefined(value.symbol)) {
-        return false;
-      }
-
-      return true;
-    })
+export const onlyAssetValidation = object().shape({
+  asset: assetValidation
 });
