+++ src/screens/swap/swap-form/swap-form.form.ts
@@ -1,7 +1,7 @@
 import { object, SchemaOf } from 'yup';
 
 import { AssetAmountInterface } from '../../../components/asset-amount-input/asset-amount-input';
-import { assetAmountValidation, assetValidation } from '../../../form/validation/asset-amount';
+import { assetAmountValidation, onlyAssetValidation } from '../../../form/validation/asset-amount';
 import { TokenInterface } from '../../../token/interfaces/token.interface';
 
 export const swapFormValidationSchema: SchemaOf<{
@@ -9,5 +9,5 @@ export const swapFormValidationSchema: SchemaOf<{
   outputAssets: { asset: TokenInterface };
 }> = object().shape({
   inputAssets: assetAmountValidation,
-  outputAssets: assetValidation
+  outputAssets: onlyAssetValidation
 });
